+++ src/brain/index.ts
@@ -14,7 +14,9 @@ export type { ActionRegistry, ActionEditor, ActionEditorContext } from './Action
 export type { SlotRegistry, SlotRenderer, SlotContext } from './SlotRegistry'
 export { SLOT_NODE_EDITOR_CHOICES } from './SlotRegistry'
 export { effectiveKind, endLabel } from './utils/nodeKind'
-export { nodeTitle, textLines, DEFAULT_NODE_TITLES } from './utils/nodeView'
+export { nodeTitle, textLines } from './utils/nodeView'
+export { NODE_KINDS, EDGE_KINDS } from './kinds'
+export type { NodeKindDescriptor, EdgeKindDescriptor, BadgeMark } from './kinds'
 export { useOpenBook, useBooks } from './hooks'
 export {
 	createBrain,